requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.2
Flask-WTF==0.14.3 # forms handling
Jinja2==2.11.2
requests==2.23.0
six==1.15.0 # Python 2 backwards compatibilty if needed
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
Werkzeug==1.0.1 # security
WTForms==2.3.1
Flask-Misaka==1.0.0 # markdown display
urllib3==1.25.9
Flask-Bootstrap4==4.0.2 # css
email_validator==1.1.1
Flask-Moment==0.9.0 # date / time formatting
